pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = 'FastapiTemplate'
version = '0.1.0'
description = 'FastapiTemplate'
requires-python = '>=3.11'
readme = "README.md"
authors = [
    {name = 'Marcel Stiebing'}
]
dependencies = [
    'uvicorn==0.21.1',
    'python-dotenv==1.0.0',
    'fastapi==0.95.1',
    'pydantic==1.10.7',
    'starlette-session==0.4.3',
    'sqlalchemy==2.0.9',
    'asyncpg==0.27.0',
]

[project.optional-dependencies]
dev = [
    'isort==5.12.0',
    'black==23.3.0',
    'flake8==6.0.0',
    'flake8-pyproject==1.2.3',
    'pytest==7.3.1',
    'httpx==0.24.0',
]

[project.urls]
'repository' = 'https://github.com/Gandori/FastapiTemplate'

[tool.black]
line-length = 88
skip-string-normalization = true

[tool.flake8]
max-line-length = 88

[tool.isort]
profile = "black"
